AM-1017: Code refactoring and clean up.
