+++ src/main/java/uk/gov/hmcts/reform/roleassignment/config/DroolConfig.java
@@ -1,7 +1,6 @@
 package uk.gov.hmcts.reform.roleassignment.config;
 
 import org.kie.api.KieServices;
-
 import org.kie.api.runtime.KieContainer;
 import org.kie.api.runtime.StatelessKieSession;
 import org.springframework.context.annotation.Bean;
@@ -10,11 +9,11 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class DroolConfig {
 
-    private KieServices  kieServices = KieServices.Factory.get();
+    private KieServices kieServices = KieServices.Factory.get();
 
     @Bean
     public KieContainer kieContainer() {
-        return  kieServices.getKieClasspathContainer();
+        return kieServices.getKieClasspathContainer();
     }
 
     @Bean
