+++ src/smokeTest/java/uk/gov/hmcts/reform/roleassignment/SmokeTest.java
@@ -17,7 +17,6 @@ import org.apache.commons.io.IOUtils;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
-import uk.gov.hmcts.reform.roleassignment.config.UserTokenProviderConfig;
 import uk.gov.hmcts.reform.roleassignment.v1.V1;
 
 import java.io.IOException;
